requirements.txt
numpy==1.24.3
absl-py==1.0.0
accelerate==0.21.0
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==0.5.2
altair==5.0.1
antlr4-python3-runtime==4.8
astropy==5.0.4
asttokens==2.1.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
blis==0.7.9
braceexpand==0.1.7
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
chex==0.1.6
click==8.1.3
cmake==3.26.4
confection==0.1.0
contextlib2==21.6.0
contexttimer==0.3.3
contourpy==1.1.0
-e git+https://github.com/facebookresearch/co-tracker.git@ee7cea937a6967009999b4b196afe91ea3fee739#egg=cotracker
cycler==0.11.0
cymem==2.0.7
decorator==4.4.2
decord==0.6.0
diffusers==0.19.0
dill==0.3.6
distlib==0.3.6
dm-haiku==0.0.9
dm-tree==0.1.8
docopt==0.6.2
easydict==1.9
EasyProcess==1.1
einops==0.4.1
einshape==1.0
entrypoint2==1.1
executing==1.2.0
fairscale==0.4.4
ffmpeg-python==0.2.0
filelock==3.12.2
filterpy==1.4.5
fire==0.4.0
flatbuffers==23.5.26
flow-vis==0.1
fonttools==4.40.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
-e git+https://github.com/IDEA-Research/Grounded-Segment-Anything.git@4689dea696cfbb2c2a359ea663faec27cf7345a4#egg=groundingdino&subdirectory=GroundingDINO
grpcio==1.56.0
h5py==3.9.0
httplib2==0.22.0
huggingface-hub==0.15.1
hydra-core==1.1.0
identify==2.5.24
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
imgaug==0.4.0
importlib-metadata==6.7.0
iopath==0.1.10
ipython==8.14.0
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jmp==0.0.4
joblib==1.3.1
jpeg4py==0.1.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
kaggle==1.5.15
keras==2.13.1
kiwisolver==1.4.4
klepto==0.2.3
kornia==0.5.0
langcodes==3.3.0
lap==0.4.0
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.9.0
lit==16.0.6
llvmlite==0.40.1
lmdb==1.4.1
loguru==0.7.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapy==1.1.8
ml-collections==0.1.1
ml-dtypes==0.2.0
more-itertools==10.0.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
murmurhash==1.0.9
natsort==8.4.0
networkx==3.1
ninja==1.11.1
nodeenv==1.8.0
numba
numpngw==0.1.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.2.1.16
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.2.60
nvidia-cuda-nvcc-cu12==12.2.91
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.2.53
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.8.15
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.5.0.53
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.1.53
nvidia-nccl-cu11==2.14.3
nvidia-nvjitlink-cu12==12.2.91
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.1.2
openai-whisper
opencv-contrib-python==4.8.0.74
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
opt-einsum==3.3.0
optax==0.1.5
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.8.0
plotly==5.15.0
portalocker==2.7.0
pox==0.3.2
pre-commit==3.3.3
preshed==3.0.8
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.11
pydeck==0.8.1b0
pyerfa==2.0.0.3
PyGetWindow==0.0.9
Pygments==2.15.1
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.1.0
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.1
python3-xlib==0.15
pytorch-lightning==1.2.9
pytweening==1.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.8.8
rsa==4.9
safetensors==0.3.1
salesforce-lavis==1.0.2
scikit-image==0.21.0
scikit-learn==1.1.1
scipy==1.11.1
seaborn==0.12.2
-e git+https://github.com/IDEA-Research/Grounded-Segment-Anything.git@4689dea696cfbb2c2a359ea663faec27cf7345a4#egg=segment_anything&subdirectory=segment_anything
sentencepiece==0.1.99
sentry-sdk==1.27.0
shapely==2.0.1
six==1.16.0
sk-video==1.1.10
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
streamlit==1.24.0
supervision==0.11.1
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==1.1.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.7.4
tiktoken==0.3.1
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.2.0
torchvision==0.15.2
tornado==6.3.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.26.1
Tree==0.2.4
triton==2.0.0
typer==0.9.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.126
uritemplate==4.1.1
urllib3==1.26.12
validators==0.20.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.5
webdataset==0.2.48
webencodings==0.5.1
Werkzeug==2.1.1
win-inet-pton==1.1.0
wincertstore==0.2
wldhx.yadisk-direct==0.0.6
wrapt==1.14.0
yacs==0.1.8
yapf==0.40.1
yarg==0.1.9
yarl==1.9.2
zipp==3.8.0
